+++ frontend/src/App.js
@@ -1,15 +1,11 @@
 import React, { Fragment , Component} from "react";
-import logo from './logo.svg';
 
 import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import Button from 'react-bootstrap/Button';
-import InputGroup  from 'react-bootstrap/InputGroup';
-import InputGroupAppend  from 'react-bootstrap/InputGroup';
 import Form from 'react-bootstrap/Form'
 import ListGroup from 'react-bootstrap/ListGroup'
 
-require
+
 /*class App extends Component{
   constructor(props) {
       super(props);
@@ -37,13 +33,11 @@ require
   }
 }*/
 
-import textSend from './textSend'
-
+import TextSend from './TextSend'
+import MessagesBox from './MessagesBox'
 function App() {  
 
-  function sayHello() {
-    console.log('Hello!');
-  }
+  
 
   return (
     
@@ -51,16 +45,9 @@ function App() {
       <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css"/>
       <div className="center">
           <Form id = "textMessagingArea">
-              <ListGroup >
-                <ListGroup.Item>Brian: Spaghetti</ListGroup.Item>
-                <ListGroup.Item>Max: Putin Putout</ListGroup.Item>
-                <ListGroup.Item>Shan: I missed my math midterm</ListGroup.Item>
-                <ListGroup.Item>Vincent: welcome to the farmer's market</ListGroup.Item>
-                <ListGroup.Item>Gideon: lol c++ embedded libraries</ListGroup.Item>
-                
-              </ListGroup>
+              <MessagesBox/>
               
-                
+                <TextSend/>
               
           </Form>
       </div>
