+++ frontend/src/textSend.js
@@ -1,16 +1,13 @@
 import React, { Fragment , Component} from "react";
-import logo from './logo.svg';
 
 import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Button from 'react-bootstrap/Button';
-import InputGroup  from 'react-bootstrap/InputGroup';
-import InputGroupAppend  from 'react-bootstrap/InputGroup';
-import Form from 'react-bootstrap/Form'
-import ListGroup from 'react-bootstrap/ListGroup'
 
-
-class textSend extends React.Component{
+function sayHello() {
+  console.log('Hello!');
+}
+class TextSend extends React.Component{
 
   render() {
       return (
@@ -23,4 +20,4 @@ class textSend extends React.Component{
 
 }
 
-export default MyComp;
+export default TextSend;
